mcp-data-service/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
google-cloud-secret-manager
google-auth